+++ src/main/java/go/alarm/service/GroupServiceImpl.java
@@ -8,8 +8,8 @@ import go.alarm.domain.repository.UserRepository;
 import go.alarm.web.converter.GroupConverter;
 import go.alarm.web.converter.UserGroupConverter;
 import go.alarm.web.converter.dayOfWeekConverter;
-import go.alarm.web.dto.GroupRequestDTO;
-import go.alarm.web.dto.GroupRequestDTO.UpdateGroupDTO;
+import go.alarm.web.dto.request.GroupRequestDTO;
+import go.alarm.web.dto.request.GroupRequestDTO.UpdateGroupDTO;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
