+++ src/main/java/go/alarm/service/UserGroupServiceImpl.java
@@ -3,14 +3,13 @@ package go.alarm.service;
 import go.alarm.domain.entity.Group;
 import go.alarm.domain.entity.User;
 import go.alarm.domain.entity.UserGroup;
-import go.alarm.domain.repository.GroupRepository;
 import go.alarm.domain.repository.UserGroupRepository;
 import go.alarm.domain.repository.UserRepository;
 import go.alarm.web.converter.MainConverter;
-import go.alarm.web.dto.MainResponseDTO;
-import go.alarm.web.dto.MainResponseDTO.GroupDTO;
-import go.alarm.web.dto.MainResponseDTO.MainDTO;
-import go.alarm.web.dto.MainResponseDTO.UserDTO;
+import go.alarm.web.dto.response.MainResponseDTO;
+import go.alarm.web.dto.response.MainResponseDTO.GroupDTO;
+import go.alarm.web.dto.response.MainResponseDTO.MainDTO;
+import go.alarm.web.dto.response.MainResponseDTO.UserDTO;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
