+++ src/main/java/go/alarm/service/UserServiceImpl.java
@@ -5,9 +5,7 @@ import go.alarm.domain.entity.User;
 import go.alarm.domain.repository.DayOfWeekRepository;
 import go.alarm.domain.repository.UserRepository;
 import go.alarm.web.converter.dayOfWeekConverter;
-import go.alarm.web.dto.UserRequestDTO;
-import go.alarm.web.dto.UserRequestDTO.SetBedTimeDTO;
-import go.alarm.web.dto.UserResponseDTO;
+import go.alarm.web.dto.request.UserRequestDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
