+++ src/main/java/go/alarm/web/controller/GroupController.java
@@ -6,8 +6,8 @@ import go.alarm.domain.entity.UserGroup;
 import go.alarm.response.BaseResponse;
 import go.alarm.service.GroupService;
 import go.alarm.web.converter.GroupConverter;
-import go.alarm.web.dto.GroupRequestDTO;
-import go.alarm.web.dto.GroupResponseDTO;
+import go.alarm.web.dto.request.GroupRequestDTO;
+import go.alarm.web.dto.response.GroupResponseDTO;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.Parameters;
