+++ src/main/java/go/alarm/web/controller/UserController.java
@@ -3,11 +3,9 @@ package go.alarm.web.controller;
 import go.alarm.domain.entity.User;
 import go.alarm.response.BaseResponse;
 import go.alarm.service.UserService;
-import go.alarm.web.converter.MyPageConverter;
 import go.alarm.web.converter.UserConverter;
-import go.alarm.web.dto.MyPageResponseDTO.myPageDTO;
-import go.alarm.web.dto.UserRequestDTO;
-import go.alarm.web.dto.UserResponseDTO;
+import go.alarm.web.dto.request.UserRequestDTO;
+import go.alarm.web.dto.response.UserResponseDTO;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.Parameters;
