+++ src/main/java/go/alarm/web/dto/request/GroupRequestDTO.java
@@ -1,4 +1,4 @@
-package go.alarm.web.dto;
+package go.alarm.web.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
